+++ src/solverpy/benchmark/db/providers/solved.py
@@ -7,16 +7,35 @@ from ...path import bids, sids
 logger = logging.getLogger(__name__)
 
 NAME = "solved"
-
+   
+def delfix(problem : str, fix : (str|int|None)) -> str:
+   """Delete a prefix of a problem name."""
+   if not fix: # covers None, 0, "" (also False)
+      return problem
+   if (type(fix) is str) and problem.startswith(fix):
+      return problem[len(fix):]
+   if (type(fix) is int) and problem.count("/") >= fix:
+      parts = problem.split("/")
+      return "/".join(parts[fix:])
+   raise Exception(f"Uknown delfix value type {type(fix)} of '{fix}'")
+      
 class Solved(CachedProvider):
 
-   def __init__(self, bid, sid, limit, store_cached=True):
+   def __init__(self, bid, sid, limit, store_cached=True, delfix=None):
       CachedProvider.__init__(self, bid, sid, limit, store_cached)
+      self._delfix = delfix
+
+   @staticmethod
+   def Maker(delfix):
+      def maker(bid, sid, limit): 
+         return Solved(bid, sid, limit, delfix=delfix)
+      return maker
 
    def store(self, task, result):
       if task.solver.solved(result):
-         if task.problem not in self.cache:
-            self.cache.add(task.problem)
+         problem = delfix(task.problem, self._delfix)
+         if problem not in self.cache:
+            self.cache.add(problem)
             self._uptodate = False
 
    def load(self):
@@ -36,4 +55,5 @@ class Solved(CachedProvider):
    def cachedump(self, fw):
       if self.cache:
          fw.write("\n".join(sorted(self.cache))+"\n")
+